updated class structure
